refactor _computeTWAP into utils.js
